+++ src/components/page-viewer/PageViewer.tsx
@@ -3,50 +3,7 @@ import PageScroller, { ScrollPosition } from './PageScroller'
 import { RefObject, useMemo, useRef, useState } from 'react'
 import ImgWrapper from './ImgWrapper'
 import { PinchEvent, usePinchPan } from '@/hooks/pinch-pan/use-pinch-pan'
-
-function clampValue(value: number, limits: { min: number; max: number }) {
-  const minBound = Math.max(limits.min, value) // prevent value from dipping below the min limit
-  return Math.min(minBound, limits.max) // prevent value from going past the max limit
-}
-
-function getScrollLimits(content: Dimensions, container: Dimensions) {
-  return {
-    top: {
-      min: 0,
-      max: content.height - container.height,
-    },
-
-    left: {
-      min: 0,
-      max: content.width - container.width,
-    },
-  }
-}
-
-function useScrolling(pageDims: Dimensions, containerDims: Dimensions) {
-  const [scroll, setScroll] = useState<ScrollPosition>({
-    left: 0,
-    top: 0,
-  })
-
-  const scrollLimits = useMemo(
-    () => getScrollLimits(pageDims, containerDims),
-    [pageDims, containerDims]
-  )
-
-  return {
-    scroll,
-    setScroll(callback: (position: ScrollPosition) => ScrollPosition) {
-      setScroll((value) => {
-        const raw = callback(value)
-        return {
-          top: clampValue(raw.top, scrollLimits.top),
-          left: clampValue(raw.left, scrollLimits.left),
-        }
-      })
-    },
-  }
-}
+import { useScrollingManager } from './use-scrolling-manager'
 
 const MAX_SCALE = 10
 const MIN_SCALE = 1
@@ -126,7 +83,7 @@ function usePinchPanInterface(
   pageDims: Dimensions,
   containerDims: Dimensions
 ) {
-  const { scroll, setScroll } = useScrolling(pageDims, containerDims)
+  const { scroll, setScroll } = useScrollingManager(pageDims, containerDims)
   const { handlePinch, scale } = usePinching(scroll, setScroll, pageDims)
 
   usePinchPan(
